refactor: split main file
