+++ main.go
@@ -22,1086 +22,7 @@ import (
 	"github.com/hajimehoshi/ebiten/v2/vector"
 )
 
-var helpMessage string
-
-func init() {
-	lines := [][2]string{
-		{"Arrow keys/WASD", "pan the camera"},
-		{"Mouse wheel or +/-", "zoom in and out"},
-		{"Drag with the mouse/touch", "pan"},
-		{"Pinch with two fingers", "zoom on touch"},
-		{"Click or tap geysers/POIs", "center and show details"},
-		{"Tap legend entries", "highlight items"},
-		{"Camera icon", "open screenshot menu"},
-		{"Geyser-icon", "list all geysers"},
-		{"Question mark", "toggle this help"},
-		{"X button", "close this help"},
-		{"Gear icon", "open options"},
-	}
-	width := 0
-	for _, p := range lines {
-		if len(p[0]) > width {
-			width = len(p[0])
-		}
-	}
-	var b strings.Builder
-	b.WriteString("Controls:\n")
-	for i, p := range lines {
-		fmt.Fprintf(&b, "%-*s | %s", width, p[0], p[1])
-		if i < len(lines)-1 {
-			b.WriteByte('\n')
-		}
-	}
-	helpMessage = b.String()
-}
-
-type hoverIcon int
-
-const (
-	hoverNone hoverIcon = iota
-	hoverOptions
-	hoverGeysers
-	hoverScreenshot
-	hoverHelp
-)
-
-var errorBorderColor = color.RGBA{R: 244, G: 67, B: 54, A: 255}
-
-var grayColorM = func() ebiten.ColorM {
-	var m ebiten.ColorM
-	const r = 0.299
-	const g = 0.587
-	const b = 0.114
-	m.SetElement(0, 0, r)
-	m.SetElement(0, 1, g)
-	m.SetElement(0, 2, b)
-	m.SetElement(1, 0, r)
-	m.SetElement(1, 1, g)
-	m.SetElement(1, 2, b)
-	m.SetElement(2, 0, r)
-	m.SetElement(2, 1, g)
-	m.SetElement(2, 2, b)
-	m.SetElement(3, 3, 1)
-	return m
-}()
-
-func drawTextWithBG(dst *ebiten.Image, text string, x, y int) {
-	w, h := textDimensions(text)
-	vector.DrawFilledRect(dst, float32(x-2), float32(y-2), float32(w+4), float32(h+4), color.RGBA{0, 0, 0, 128}, false)
-	drawText(dst, text, x, y)
-}
-
-func drawTextWithBGScale(dst *ebiten.Image, text string, x, y int, scale float64) {
-	if scale == 1.0 {
-		drawTextWithBG(dst, text, x, y)
-		return
-	}
-	w, h := textDimensions(text)
-	w += 4
-	h += 4
-	img := ebiten.NewImage(w, h)
-	vector.DrawFilledRect(img, 0, 0, float32(w), float32(h), color.RGBA{0, 0, 0, 128}, false)
-	drawText(img, text, 2, 2)
-	op := &ebiten.DrawImageOptions{}
-	op.GeoM.Scale(scale, scale)
-	op.GeoM.Translate(float64(x-2), float64(y-2))
-	dst.DrawImage(img, op)
-}
-
-func drawTextWithBGBorder(dst *ebiten.Image, text string, x, y int, border color.Color) {
-	w, h := textDimensions(text)
-	bx := x - 2
-	by := y - 2
-	bw := w + 4
-	bh := h + 4
-	vector.DrawFilledRect(dst, float32(bx-1), float32(by-1), float32(bw+2), float32(bh+2), border, false)
-	vector.DrawFilledRect(dst, float32(bx), float32(by), float32(bw), float32(bh), color.RGBA{0, 0, 0, 128}, false)
-	drawText(dst, text, x, y)
-}
-
-func drawTextWithBGBorderScale(dst *ebiten.Image, text string, x, y int, border color.Color, scale float64) {
-	if scale == 1.0 {
-		drawTextWithBGBorder(dst, text, x, y, border)
-		return
-	}
-	w, h := textDimensions(text)
-	w += 4
-	h += 4
-	img := ebiten.NewImage(w+2, h+2)
-	vector.DrawFilledRect(img, 0, 0, float32(w+2), float32(h+2), border, false)
-	vector.DrawFilledRect(img, 1, 1, float32(w), float32(h), color.RGBA{0, 0, 0, 128}, false)
-	drawText(img, text, 3, 3)
-	op := &ebiten.DrawImageOptions{}
-	op.GeoM.Scale(scale, scale)
-	op.GeoM.Translate(float64(x-2), float64(y-2))
-	dst.DrawImage(img, op)
-}
-
-func drawTextScale(dst *ebiten.Image, text string, x, y int, scale float64) {
-	if scale == 1.0 {
-		drawText(dst, text, x, y)
-		return
-	}
-	w, h := textDimensions(text)
-	img := ebiten.NewImage(w, h)
-	drawText(img, text, 0, 0)
-	op := &ebiten.DrawImageOptions{}
-	op.GeoM.Scale(scale, scale)
-	op.GeoM.Translate(float64(x), float64(y))
-	dst.DrawImage(img, op)
-}
-
-func (g *Game) drawInfoPanel(dst *ebiten.Image, text string, icon *ebiten.Image, x, y int, scale float64) {
-	txtW, txtH := textDimensions(text)
-	iconW, iconH := 0, 0
-	if icon != nil {
-		iconW = InfoIconSize
-		iconH = InfoIconSize
-	}
-	gap := 4
-	w := txtW + iconW + gap + 8
-	h := txtH
-	if iconH > txtH {
-		h = iconH
-	}
-	h += 8
-	img := ebiten.NewImage(w, h)
-	vector.DrawFilledRect(img, 0, 0, float32(w), float32(h), color.RGBA{0, 0, 0, InfoPanelAlpha}, false)
-	vector.StrokeRect(img, 0.5, 0.5, float32(w)-1, float32(h)-1, 1, buttonBorderColor, false)
-	if icon != nil {
-		opIcon := &ebiten.DrawImageOptions{Filter: g.filterMode()}
-		scaleIcon := float64(InfoIconSize) / math.Max(float64(icon.Bounds().Dx()), float64(icon.Bounds().Dy()))
-		opIcon.GeoM.Scale(scaleIcon, scaleIcon)
-		opIcon.GeoM.Translate(4, float64(h-iconH)/2)
-		img.DrawImage(icon, opIcon)
-	}
-	drawText(img, text, iconW+gap+4, 4)
-	op := &ebiten.DrawImageOptions{}
-	op.GeoM.Scale(scale, scale)
-	op.GeoM.Translate(float64(x-4), float64(y-4))
-	dst.DrawImage(img, op)
-}
-
-func (g *Game) drawInfoRow(dst *ebiten.Image, text string, icon *ebiten.Image, x, y int, scale float64) {
-	txtW, txtH := textDimensions(text)
-	iconW, iconH := 0, 0
-	if icon != nil {
-		iconW = InfoIconSize
-		iconH = InfoIconSize
-	}
-	gap := 4
-	w := txtW + iconW + gap
-	h := txtH
-	if iconH > txtH {
-		h = iconH
-	}
-	if scale == 1.0 {
-		if icon != nil {
-			opIcon := &ebiten.DrawImageOptions{Filter: g.filterMode()}
-			sc := float64(InfoIconSize) / math.Max(float64(icon.Bounds().Dx()), float64(icon.Bounds().Dy()))
-			opIcon.GeoM.Scale(sc, sc)
-			opIcon.GeoM.Translate(float64(x), float64(y+(h-iconH)/2))
-			dst.DrawImage(icon, opIcon)
-		}
-		drawText(dst, text, x+iconW+gap, y+(h-txtH)/2)
-		return
-	}
-
-	img := ebiten.NewImage(w, h)
-	if icon != nil {
-		opIcon := &ebiten.DrawImageOptions{Filter: g.filterMode()}
-		sc := float64(InfoIconSize) / math.Max(float64(icon.Bounds().Dx()), float64(icon.Bounds().Dy()))
-		opIcon.GeoM.Scale(sc, sc)
-		opIcon.GeoM.Translate(0, float64(h-iconH)/2)
-		img.DrawImage(icon, opIcon)
-	}
-	drawText(img, text, iconW+gap, (h-txtH)/2)
-	op := &ebiten.DrawImageOptions{}
-	op.GeoM.Scale(scale, scale)
-	op.GeoM.Translate(float64(x), float64(y))
-	dst.DrawImage(img, op)
-}
-
-func infoPanelSize(text string, icon *ebiten.Image) (int, int) {
-	txtW, txtH := textDimensions(text)
-	iconW, iconH := 0, 0
-	if icon != nil {
-		iconW = InfoIconSize
-		iconH = InfoIconSize
-	}
-	gap := 4
-	w := txtW + iconW + gap + 8
-	h := txtH
-	if iconH > txtH {
-		h = iconH
-	}
-	h += 8
-	return w, h
-}
-
-func infoRowSize(text string, icon *ebiten.Image) (int, int) {
-	txtW, txtH := textDimensions(text)
-	iconW, iconH := 0, 0
-	if icon != nil {
-		iconW = InfoIconSize
-		iconH = InfoIconSize
-	}
-	gap := 4
-	w := txtW + iconW + gap
-	h := txtH
-	if iconH > txtH {
-		h = iconH
-	}
-	return w, h
-}
-
-func drawBiome(dst *ebiten.Image, polys [][]Point, clr color.Color, camX, camY, zoom float64) {
-	if len(polys) == 0 {
-		return
-	}
-	var p vector.Path
-	for _, pts := range polys {
-		if len(pts) == 0 {
-			continue
-		}
-		p.MoveTo(float32(pts[0].X*2), float32(pts[0].Y*2))
-		for _, pt := range pts[1:] {
-			p.LineTo(float32(pt.X*2), float32(pt.Y*2))
-		}
-		p.Close()
-	}
-	vs, is := p.AppendVerticesAndIndicesForFilling(nil, nil)
-	r, g, b, a := clr.RGBA()
-	for i := range vs {
-		x := float64(vs[i].DstX)*zoom + camX
-		y := float64(vs[i].DstY)*zoom + camY
-		vs[i].DstX = float32(x)
-		vs[i].DstY = float32(y)
-		vs[i].SrcX = 0
-		vs[i].SrcY = 0
-		vs[i].ColorR = float32(r) / 0xffff
-		vs[i].ColorG = float32(g) / 0xffff
-		vs[i].ColorB = float32(b) / 0xffff
-		vs[i].ColorA = float32(a) / 0xffff
-	}
-	op := &ebiten.DrawTrianglesOptions{
-		AntiAlias:      true,
-		ColorScaleMode: ebiten.ColorScaleModePremultipliedAlpha,
-		FillRule:       ebiten.FillRuleEvenOdd,
-	}
-	dst.DrawTriangles(vs, is, whitePixel, op)
-}
-
-func drawBiomeTextured(dst *ebiten.Image, polys [][]Point, tex *ebiten.Image, clr color.Color, camX, camY, zoom float64, filter ebiten.Filter) {
-	if len(polys) == 0 || tex == nil {
-		return
-	}
-	var p vector.Path
-	for _, pts := range polys {
-		if len(pts) == 0 {
-			continue
-		}
-		p.MoveTo(float32(pts[0].X*2), float32(pts[0].Y*2))
-		for _, pt := range pts[1:] {
-			p.LineTo(float32(pt.X*2), float32(pt.Y*2))
-		}
-		p.Close()
-	}
-	vs, is := p.AppendVerticesAndIndicesForFilling(nil, nil)
-	r, g, b, a := clr.RGBA()
-	texW, texH := tex.Bounds().Dx(), tex.Bounds().Dy()
-	for i := range vs {
-		worldX := float64(vs[i].DstX) / 2
-		worldY := float64(vs[i].DstY) / 2
-		x := float64(vs[i].DstX)*zoom + camX
-		y := float64(vs[i].DstY)*zoom + camY
-		vs[i].DstX = float32(x)
-		vs[i].DstY = float32(y)
-		vs[i].SrcX = float32(worldX*BiomeTextureScale) * float32(texW)
-		vs[i].SrcY = float32(worldY*BiomeTextureScale) * float32(texH)
-		vs[i].ColorR = float32(r) / 0xffff
-		vs[i].ColorG = float32(g) / 0xffff
-		vs[i].ColorB = float32(b) / 0xffff
-		vs[i].ColorA = float32(a) / 0xffff
-	}
-	op := &ebiten.DrawTrianglesOptions{
-		AntiAlias:      true,
-		ColorScaleMode: ebiten.ColorScaleModePremultipliedAlpha,
-		FillRule:       ebiten.FillRuleEvenOdd,
-		Address:        ebiten.AddressRepeat,
-		Filter:         filter,
-	}
-	dst.DrawTriangles(vs, is, tex, op)
-}
-
-func drawBiomeOutline(dst *ebiten.Image, polys [][]Point, camX, camY, zoom float64, clr color.Color) {
-	for _, pts := range polys {
-		if len(pts) < 2 {
-			continue
-		}
-		for i := 0; i < len(pts); i++ {
-			a := pts[i]
-			b := pts[(i+1)%len(pts)]
-			x0 := float32(math.Round(float64(a.X*2)*zoom + camX))
-			y0 := float32(math.Round(float64(a.Y*2)*zoom + camY))
-			x1 := float32(math.Round(float64(b.X*2)*zoom + camX))
-			y1 := float32(math.Round(float64(b.Y*2)*zoom + camY))
-			vector.StrokeLine(dst, x0, y0, x1, y1, 1, clr, true)
-		}
-	}
-}
-
-// buildLegendImage returns an image of biome colors with a single background.
-func buildLegendImage(biomes []BiomePath) (*ebiten.Image, []string) {
-	set := make(map[string]struct{})
-	for _, b := range biomes {
-		set[b.Name] = struct{}{}
-	}
-	names := make([]string, 0, len(set))
-	for n := range set {
-		names = append(names, n)
-	}
-	sort.Strings(names)
-
-	maxW := 0
-	for _, name := range names {
-		w, _ := textDimensions(displayBiome(name))
-		if w > maxW {
-			maxW = w
-		}
-	}
-
-	spacing := rowSpacing()
-	width := 30 + maxW + 5
-	height := spacing*(len(names)+2) + 7
-
-	img := ebiten.NewImage(width, height)
-	img.Fill(color.RGBA{0, 0, 0, 77})
-
-	y := 10
-	drawTextWithBG(img, "Biomes", 5, y)
-	y += spacing
-	for _, name := range names {
-		clr, ok := biomeColors[name]
-		if !ok {
-			clr = color.RGBA{60, 60, 60, 255}
-		}
-		vector.DrawFilledRect(img, 5, float32(y), 20, 10, clr, false)
-		drawTextWithBGBorder(img, displayBiome(name), 30, y, clr)
-		y += spacing
-	}
-
-	drawTextWithBGBorder(img, "Clear", 5, y, buttonBorderColor)
-
-	return img, names
-}
-
-// initObjectLegend prepares the mapping of object names to numeric labels and
-// caches the legend text.
-func (g *Game) initObjectLegend() {
-	if g.legendMap != nil {
-		return
-	}
-	g.legendMap = make(map[string]int)
-	g.legendColors = nil
-	counter := 1
-	for _, gy := range g.geysers {
-		name := displayGeyser(gy.ID)
-		if _, ok := g.legendMap["g"+name]; !ok {
-			g.legendMap["g"+name] = counter
-			g.legendEntries = append(g.legendEntries, fmt.Sprintf("%d: %s", counter, name))
-			g.legendColors = append(g.legendColors, uniqueColor(counter))
-			counter++
-		}
-	}
-	for _, poi := range g.pois {
-		name := displayPOI(poi.ID)
-		if _, ok := g.legendMap["p"+name]; !ok {
-			g.legendMap["p"+name] = counter
-			g.legendEntries = append(g.legendEntries, fmt.Sprintf("%d: %s", counter, name))
-			g.legendColors = append(g.legendColors, uniqueColor(counter))
-			counter++
-		}
-	}
-}
-
-func (g *Game) invalidateLegends() {
-	g.legend = nil
-	g.legendImage = nil
-}
-
-// drawNumberLegend draws the cached legend image on the top right corner with
-// a single semi-transparent background.
-func (g *Game) drawNumberLegend(dst *ebiten.Image) {
-	if len(g.legendEntries) == 0 {
-		return
-	}
-	if g.legendImage == nil {
-		maxW := 0
-		for _, e := range g.legendEntries {
-			w, _ := textDimensions(e)
-			if w > maxW {
-				maxW = w
-			}
-		}
-		spacing := rowSpacing()
-		width := maxW + 10
-		height := spacing*(len(g.legendEntries)+2) + 7
-		img := ebiten.NewImage(width, height)
-		img.Fill(color.RGBA{0, 0, 0, 77})
-		y := 10
-		drawTextWithBG(img, "Items", 5, y)
-		y += spacing
-		for i, e := range g.legendEntries {
-			clr := color.RGBA{}
-			if i < len(g.legendColors) {
-				clr = g.legendColors[i]
-			}
-			drawTextWithBGBorder(img, e, 5, y, clr)
-			y += spacing
-		}
-		drawTextWithBGBorder(img, "Clear", 5, y, buttonBorderColor)
-		g.legendImage = img
-	}
-	scale := g.uiScale()
-	w := float64(g.legendImage.Bounds().Dx()) * scale
-	x := float64(dst.Bounds().Dx()) - w - 12
-	y := 10.0 - g.itemScroll
-	op := &ebiten.DrawImageOptions{}
-	op.GeoM.Scale(scale, scale)
-	op.GeoM.Translate(math.Round(x), math.Round(y))
-	dst.DrawImage(g.legendImage, op)
-	lh := float64(g.legendImage.Bounds().Dy()) * scale
-	strokeFrame(dst, image.Rect(int(math.Round(x)), int(math.Round(y)), int(math.Round(x+w)), int(math.Round(y+lh))))
-	if g.selectedItem >= 0 {
-		spacing := float64(rowSpacing())
-		hy := y + (10+spacing*float64(g.selectedItem+1))*scale
-		hh := spacing * scale
-		vector.StrokeRect(dst, float32(math.Round(x))+0.5, float32(math.Round(hy))-4, float32(math.Round(w))-1, float32(math.Round(hh))-1, 2, color.RGBA{255, 0, 0, 255}, false)
-	}
-}
-
-func (g *Game) drawGeyserList(dst *ebiten.Image) {
-	vector.DrawFilledRect(dst, 0, 0, float32(g.width), float32(g.height), color.RGBA{0, 0, 0, 255}, false)
-	scale := g.uiScale()
-	cr := g.geyserCloseRect()
-	drawCloseButton(dst, cr)
-
-	spacing := int(10 * scale)
-	type item struct {
-		text string
-		icon *ebiten.Image
-		w    int
-		h    int
-	}
-	items := make([]item, len(g.geysers))
-	maxW := 0
-	for i, gy := range g.geysers {
-		ic := (*ebiten.Image)(nil)
-		if n := iconForGeyser(gy.ID); n != "" {
-			ic = g.icons[n]
-		}
-		txt := displayGeyser(gy.ID) + "\n" + formatGeyserInfo(gy)
-		w, h := infoRowSize(txt, ic)
-		sw := int(float64(w) * scale)
-		sh := int(float64(h) * scale)
-		if sw > maxW {
-			maxW = sw
-		}
-		items[i] = item{text: txt, icon: ic, w: sw, h: sh}
-	}
-	cols := 1
-	if maxW+spacing > 0 {
-		cols = g.width / (maxW + spacing)
-		if cols < 1 {
-			cols = 1
-		}
-	}
-	rows := (len(items) + cols - 1) / cols
-	rowHeights := make([]int, rows)
-	for i, it := range items {
-		r := i / cols
-		if it.h > rowHeights[r] {
-			rowHeights[r] = it.h
-		}
-	}
-	y := spacing - int(g.geyserScroll)
-	idx := 0
-	for r := 0; r < rows; r++ {
-		x := spacing
-		for c := 0; c < cols && idx < len(items); c++ {
-			it := items[idx]
-			g.drawInfoRow(dst, it.text, it.icon, x, y, scale)
-			x += maxW + spacing
-			idx++
-		}
-		y += rowHeights[r] + spacing
-	}
-}
-
-func (g *Game) maxGeyserScroll() float64 {
-	scale := g.uiScale()
-	spacing := int(10 * scale)
-	type item struct {
-		w int
-		h int
-	}
-	items := make([]item, len(g.geysers))
-	maxW := 0
-	for i, gy := range g.geysers {
-		ic := (*ebiten.Image)(nil)
-		if n := iconForGeyser(gy.ID); n != "" {
-			ic = g.icons[n]
-		}
-		txt := displayGeyser(gy.ID) + "\n" + formatGeyserInfo(gy)
-		w, h := infoRowSize(txt, ic)
-		sw := int(float64(w) * scale)
-		sh := int(float64(h) * scale)
-		if sw > maxW {
-			maxW = sw
-		}
-		items[i] = item{w: sw, h: sh}
-	}
-	cols := 1
-	if maxW+spacing > 0 {
-		cols = g.width / (maxW + spacing)
-		if cols < 1 {
-			cols = 1
-		}
-	}
-	rows := (len(items) + cols - 1) / cols
-	if rows == 0 {
-		return 0
-	}
-	rowHeights := make([]int, rows)
-	for i, it := range items {
-		r := i / cols
-		if it.h > rowHeights[r] {
-			rowHeights[r] = it.h
-		}
-	}
-	total := spacing
-	for _, h := range rowHeights {
-		total += h + spacing
-	}
-	max := total - rowHeights[rows-1] - spacing
-	if max < 0 {
-		max = 0
-	}
-	return float64(max)
-}
-
-func (g *Game) adjustGeyserScroll(delta float64) {
-	g.geyserScroll += delta
-	if g.geyserScroll < 0 {
-		g.geyserScroll = 0
-	}
-	if max := g.maxGeyserScroll(); g.geyserScroll > max {
-		g.geyserScroll = max
-	}
-	g.needsRedraw = true
-}
-
-func (g *Game) maxBiomeScroll() float64 {
-	if g.legend == nil {
-		return 0
-	}
-	scale := g.uiScale()
-	h := float64(g.legend.Bounds().Dy()) * scale
-	extra := float64(rowSpacing()*LegendScrollExtraRows) * scale
-	max := h - float64(g.height) + extra
-	if max < 0 {
-		max = 0
-	}
-	return max
-}
-
-func (g *Game) maxItemScroll() float64 {
-	if g.legendImage == nil {
-		return 0
-	}
-	scale := g.uiScale()
-	h := float64(g.legendImage.Bounds().Dy()) * scale
-	extra := float64(rowSpacing()*LegendScrollExtraRows) * scale
-	max := h + 10 - float64(g.height) + extra
-	if max < 0 {
-		max = 0
-	}
-	return max
-}
-
-func (g *Game) itemPanelVisible() bool {
-	return g.showLegend && g.useNumbers && g.showItemNames && g.zoom < LegendZoomThreshold && !g.screenshotMode
-}
-
-func (g *Game) updateHover(mx, my int) {
-	prevBiome := g.hoverBiome
-	prevItem := g.hoverItem
-	g.hoverBiome = -1
-	g.hoverItem = -1
-	if !g.showLegend {
-		if prevBiome != -1 || prevItem != -1 {
-			g.needsRedraw = true
-		}
-		return
-	}
-	scale := g.uiScale()
-	if g.legend != nil && len(g.legendBiomes) > 0 {
-		w := int(float64(g.legend.Bounds().Dx()) * scale)
-		if mx >= 0 && mx < w {
-			spacing := float64(rowSpacing())
-			baseY := int((10+spacing)*scale) - int(g.biomeScroll)
-			rowH := int(spacing * scale)
-			for i := range g.legendBiomes {
-				ry := baseY + i*rowH
-				if my >= ry && my < ry+rowH {
-					g.hoverBiome = i
-					break
-				}
-			}
-			if g.hoverBiome == -1 {
-				ry := baseY + len(g.legendBiomes)*rowH
-				if my >= ry && my < ry+rowH {
-					g.hoverBiome = len(g.legendBiomes)
-				}
-			}
-		}
-	}
-	useNumbers := g.useNumbers && g.showItemNames && g.zoom < LegendZoomThreshold && !g.screenshotMode
-	if useNumbers && g.legendImage != nil {
-		w := int(float64(g.legendImage.Bounds().Dx()) * scale)
-		x0 := g.width - w - 12
-		if mx >= x0 && mx < x0+w {
-			spacing := float64(rowSpacing())
-			baseY := int((10+spacing)*scale) - int(g.itemScroll)
-			rowH := int(spacing * scale)
-			for i := range g.legendEntries {
-				ry := baseY + i*rowH
-				if my >= ry && my < ry+rowH {
-					g.hoverItem = i
-					break
-				}
-			}
-			if g.hoverItem == -1 {
-				ry := baseY + len(g.legendEntries)*rowH
-				if my >= ry && my < ry+rowH {
-					g.hoverItem = len(g.legendEntries)
-				}
-			}
-		}
-	}
-	if g.hoverBiome != prevBiome || g.hoverItem != prevItem {
-		g.needsRedraw = true
-	}
-}
-
-func (g *Game) updateIconHover(mx, my int) {
-	prev := g.hoverIcon
-	g.hoverIcon = hoverNone
-	if mx == -1 || my == -1 {
-		if prev != g.hoverIcon {
-			g.needsRedraw = true
-		}
-		return
-	}
-	pt := image.Rect(mx, my, mx+1, my+1)
-	switch {
-	case g.optionsRect().Overlaps(pt):
-		g.hoverIcon = hoverOptions
-	case g.geyserRect().Overlaps(pt):
-		g.hoverIcon = hoverGeysers
-	case g.screenshotRect().Overlaps(pt):
-		g.hoverIcon = hoverScreenshot
-	case g.helpRect().Overlaps(pt):
-		g.hoverIcon = hoverHelp
-	}
-	if prev != g.hoverIcon {
-		g.needsRedraw = true
-	}
-}
-
-func (g *Game) clickLegend(mx, my int) bool {
-	if !g.showLegend {
-		return false
-	}
-	handled := false
-	scale := g.uiScale()
-	if g.legend != nil && len(g.legendBiomes) > 0 {
-		w := int(float64(g.legend.Bounds().Dx()) * scale)
-		if mx >= 0 && mx < w {
-			spacing := float64(rowSpacing())
-			baseY := int((10+spacing)*scale) - int(g.biomeScroll)
-			rowH := int(spacing * scale)
-			count := len(g.legendBiomes)
-			for i := 0; i <= count; i++ {
-				ry := baseY + i*rowH
-				if my >= ry && my < ry+rowH {
-					if i == count {
-						g.selectedBiome = -1
-					} else {
-						g.selectedBiome = i
-					}
-					handled = true
-					break
-				}
-			}
-		}
-	}
-	useNumbers := g.useNumbers && g.showItemNames && g.zoom < LegendZoomThreshold && !g.screenshotMode
-	if useNumbers && g.legendImage != nil {
-		w := int(float64(g.legendImage.Bounds().Dx()) * scale)
-		x0 := g.width - w - 12
-		if mx >= x0 && mx < x0+w {
-			spacing := float64(rowSpacing())
-			baseY := int((10+spacing)*scale) - int(g.itemScroll)
-			rowH := int(spacing * scale)
-			count := len(g.legendEntries)
-			for i := 0; i <= count; i++ {
-				ry := baseY + i*rowH
-				if my >= ry && my < ry+rowH {
-					if i == count {
-						g.selectedItem = -1
-					} else {
-						g.selectedItem = i
-					}
-					handled = true
-					break
-				}
-			}
-		}
-	}
-	if handled {
-		g.needsRedraw = true
-	}
-	return handled
-}
-
 // Game implements ebiten.Game and displays geysers with their names.
-type Game struct {
-	geysers           []Geyser
-	pois              []PointOfInterest
-	biomes            []BiomePath
-	asteroids         []Asteroid
-	icons             map[string]*ebiten.Image
-	biomeTextures     map[string]*ebiten.Image
-	width             int
-	height            int
-	astWidth          int
-	astHeight         int
-	camX              float64
-	camY              float64
-	zoom              float64
-	minZoom           float64
-	dragging          bool
-	lastX             int
-	lastY             int
-	touches           map[ebiten.TouchID]touchPoint
-	pinchDist         float64
-	needsRedraw       bool
-	screenshotPath    string
-	captured          bool
-	legend            *ebiten.Image
-	legendMap         map[string]int
-	legendEntries     []string
-	legendColors      []color.RGBA
-	legendImage       *ebiten.Image
-	legendBiomes      []string
-	hoverBiome        int
-	hoverItem         int
-	selectedBiome     int
-	selectedItem      int
-	showGeyserList    bool
-	geyserScroll      float64
-	biomeScroll       float64
-	itemScroll        float64
-	showHelp          bool
-	lastWheel         time.Time
-	loading           bool
-	status            string
-	iconResults       chan loadedIcon
-	coord             string
-	mobile            bool
-	showInfo          bool
-	infoPinned        bool
-	infoText          string
-	infoX             int
-	infoY             int
-	infoIcon          *ebiten.Image
-	lastMouseX        int
-	lastMouseY        int
-	hoverIcon         hoverIcon
-	touchUsed         bool
-	touchActive       bool
-	touchStartX       int
-	touchStartY       int
-	touchMoved        bool
-	touchUI           bool
-	showShotMenu      bool
-	showAstMenu       bool
-	showOptions       bool
-	asteroidScroll    float64
-	screenshotMode    bool
-	ssQuality         int
-	ssSaved           time.Time
-	ssPending         int
-	skipClickTicks    int
-	lastDraw          time.Time
-	wasMinimized      bool
-	asteroidID        string
-	asteroidSpecified bool
-	statusError       bool
-	fitOnLoad         bool
-
-	textures      bool
-	vsync         bool
-	showItemNames bool
-	showLegend    bool
-	useNumbers    bool
-	iconScale     float64
-	smartRender   bool
-	linearFilter  bool
-
-	noColor   bool
-	ssNoColor bool
-	grayImage *ebiten.Image
-}
-
-type label struct {
-	text  string
-	x     int
-	y     int
-	width int
-	clr   color.RGBA
-}
-
-type touchPoint struct {
-	x int
-	y int
-}
-
-type loadedIcon struct {
-	name string
-	img  *ebiten.Image
-}
-
-// uiScale returns a multiplier for UI elements based on the current window
-// height. Larger windows and high quality screenshots use bigger values so
-// that text and icons remain readable.
-func (g *Game) uiScale() float64 {
-	if g.mobile {
-		return 1.0
-	}
-	if g.screenshotMode {
-		if g.height > 1700 {
-			return 4.0
-		}
-		if g.height > 850 {
-			return 2.0
-		}
-		return 1.0
-	}
-	return 1.0
-}
-
-func (g *Game) iconSize() int {
-	return int(float64(HelpIconSize) * g.uiScale())
-}
-
-func (g *Game) filterMode() ebiten.Filter {
-	if g.linearFilter {
-		return ebiten.FilterLinear
-	}
-	return ebiten.FilterNearest
-}
-
-func (g *Game) inBounds(x, y int) bool {
-	return x >= 0 && x < g.width && y >= 0 && y < g.height
-}
-
-func drawPlusMinus(dst *ebiten.Image, rect image.Rectangle, minus bool) {
-	cx := float32(rect.Min.X + rect.Dx()/2)
-	cy := float32(rect.Min.Y + rect.Dy()/2)
-	length := float32(rect.Dx()) * 0.5
-	thickness := float32(rect.Dx()) / 8
-	vector.StrokeLine(dst, cx-length/2, cy, cx+length/2, cy, thickness, color.RGBA{255, 255, 255, 255}, true)
-	if !minus {
-		vector.StrokeLine(dst, cx, cy-length/2, cx, cy+length/2, thickness, color.RGBA{255, 255, 255, 255}, true)
-	}
-}
-
-func (g *Game) drawTooltip(dst *ebiten.Image, text string, rect image.Rectangle, scale float64) {
-	w, h := textDimensions(text)
-	x := rect.Min.X + rect.Dx()/2 - int(float64(w)*scale/2)
-	y := rect.Min.Y - int(float64(h)*scale) - 4
-	if x < 0 {
-		x = 0
-	}
-	if x+int(float64(w)*scale) > g.width {
-		x = g.width - int(float64(w)*scale)
-	}
-	if y < 0 {
-		y = rect.Max.Y + 4
-	}
-	drawTextWithBGScale(dst, text, x, y, scale)
-}
-
-func (g *Game) helpRect() image.Rectangle {
-	size := g.iconSize()
-	x := g.width - size - HelpMargin
-	y := g.height - size - HelpMargin
-	return image.Rect(x, y, x+size, y+size)
-}
-
-func helpMenuSize() (int, int) {
-	w, h := textDimensions(helpMessage)
-	return w + 4, h + 4
-}
-
-func (g *Game) helpMenuRect() image.Rectangle {
-	w, h := helpMenuSize()
-	scale := g.uiScale()
-	w = int(float64(w) * scale)
-	h = int(float64(h) * scale)
-	r := g.helpRect()
-	x := r.Max.X - w
-	if x < 0 {
-		x = 0
-	}
-	if x+w > g.width {
-		x = g.width - w
-	}
-	y := r.Min.Y - h
-	if y < 0 {
-		y = 0
-	}
-	return image.Rect(x, y, x+w, y+h)
-}
-
-func (g *Game) helpCloseRect() image.Rectangle {
-	size := g.iconSize()
-	r := g.helpMenuRect()
-	return image.Rect(r.Max.X-size-2, r.Min.Y+2, r.Max.X-2, r.Min.Y+size+2)
-}
-
-func (g *Game) geyserRect() image.Rectangle {
-	size := g.iconSize()
-	x := g.width - size*3 - HelpMargin*3
-	y := g.height - size - HelpMargin
-	return image.Rect(x, y, x+size, y+size)
-}
-
-func (g *Game) geyserCloseRect() image.Rectangle {
-	size := g.iconSize()
-	x := g.width - size - HelpMargin
-	y := HelpMargin
-	return image.Rect(x, y, x+size, y+size)
-}
-
-func (g *Game) bottomTrayRect() image.Rectangle {
-	r := g.optionsRect()
-	r = r.Union(g.geyserRect())
-	r = r.Union(g.screenshotRect())
-	r = r.Union(g.helpRect())
-	return image.Rect(r.Min.X-4, r.Min.Y-4, r.Max.X+4, r.Max.Y+4)
-}
-
-func (g *Game) clampCamera() {
-	if g.astWidth == 0 || g.astHeight == 0 {
-		return
-	}
-	cx := float64(g.width) / 2
-	cy := float64(g.height) / 2
-	maxX := cx
-	minX := cx - float64(g.astWidth)*2*g.zoom
-	if g.camX < minX {
-		g.camX = minX
-	}
-	if g.camX > maxX {
-		g.camX = maxX
-	}
-	maxY := cy
-	minY := cy - float64(g.astHeight)*2*g.zoom
-	if g.camY < minY {
-		g.camY = minY
-	}
-	if g.camY > maxY {
-		g.camY = maxY
-	}
-}
-
-func (g *Game) itemAt(mx, my int) (string, int, int, *ebiten.Image, bool) {
-	const hitRadius = 10
-	for _, gy := range g.geysers {
-		x := float64(gy.X)*2*g.zoom + g.camX
-		y := float64(gy.Y)*2*g.zoom + g.camY
-		left, top, right, bottom := x-hitRadius, y-hitRadius, x+hitRadius, y+hitRadius
-		if iconName := iconForGeyser(gy.ID); iconName != "" {
-			if img, ok := g.icons[iconName]; ok && img != nil {
-				maxDim := math.Max(float64(img.Bounds().Dx()), float64(img.Bounds().Dy()))
-				scale := g.zoom * IconScale * g.iconScale * float64(BaseIconPixels) / maxDim
-				w := float64(img.Bounds().Dx()) * scale
-				h := float64(img.Bounds().Dy()) * scale
-				left = x - w/2
-				top = y - h/2
-				right = x + w/2
-				bottom = y + h/2
-			}
-		}
-		if float64(mx) >= left && float64(mx) <= right && float64(my) >= top && float64(my) <= bottom {
-			info := displayGeyser(gy.ID) + "\n" + formatGeyserInfo(gy)
-			var icon *ebiten.Image
-			if n := iconForGeyser(gy.ID); n != "" {
-				icon = g.icons[n]
-			}
-			return info, int(math.Round(x)), int(math.Round(y)), icon, true
-		}
-	}
-	for _, poi := range g.pois {
-		x := float64(poi.X)*2*g.zoom + g.camX
-		y := float64(poi.Y)*2*g.zoom + g.camY
-		left, top, right, bottom := x-hitRadius, y-hitRadius, x+hitRadius, y+hitRadius
-		if iconName := iconForPOI(poi.ID); iconName != "" {
-			if img, ok := g.icons[iconName]; ok && img != nil {
-				maxDim := math.Max(float64(img.Bounds().Dx()), float64(img.Bounds().Dy()))
-				scale := g.zoom * IconScale * g.iconScale * float64(BaseIconPixels) / maxDim
-				w := float64(img.Bounds().Dx()) * scale
-				h := float64(img.Bounds().Dy()) * scale
-				left = x - w/2
-				top = y - h/2
-				right = x + w/2
-				bottom = y + h/2
-			}
-		}
-		if float64(mx) >= left && float64(mx) <= right && float64(my) >= top && float64(my) <= bottom {
-			info := displayPOI(poi.ID) + "\n" + formatPOIInfo(poi)
-			var icon *ebiten.Image
-			if n := iconForPOI(poi.ID); n != "" {
-				icon = g.icons[n]
-			}
-			return info, int(math.Round(x)), int(math.Round(y)), icon, true
-		}
-	}
-	return "", 0, 0, nil, false
-}
-
-func abs(a int) int {
-	if a < 0 {
-		return -a
-	}
-	return a
-}
-
-func (g *Game) startIconLoader(names []string) {
-	g.iconResults = make(chan loadedIcon, len(names))
-	go func() {
-		for _, n := range names {
-			img, _ := loadImageFile(n)
-			g.iconResults <- loadedIcon{name: n, img: img}
-		}
-		close(g.iconResults)
-	}()
-}
-
 func (g *Game) Update() error {
 	const panSpeed = PanSpeed
